Support inline sourcemaps with `sourcemap` option
PostCSS requires a `from` and `to` filename in order to process inline
(base64-encoded) sourcemaps, so we include the name of the file passed
to us from Broccoli.

By default, PostCSS will output an inline sourcemap if one exists in
the input, and omit one if none was initially present. The `sourcemap`
option can be set to true or false to override this behavior.